Extract has_privileges
